Implement data-confirm

+++ app/views/users/profile.tpl
@@ -13,7 +13,7 @@
   <section class="my-game-ads">
     {foreach from=$game_ads item=ad}
       <h1>Game ad: {$ad->game->name} ({$ad->console->name})</h1>
-      <a href="{$site_root}/ads/destroy?id={$ad->id}">Rimuovi</a>
+      <a data-confirm href="{$site_root}/ads/destroy?id={$ad->id}">Rimuovi</a>
       <a href="{$site_root}/ads/edit?id={$ad->id}">Modifica</a>
     {/foreach}
   </section>
@@ -21,7 +21,7 @@
   <section class="my-accessory-ads">
     {foreach from=$accessory_ads item=ad}
       <h1>Accessory ad: {$ad->accessory->name} ({$ad->console->name})</h1>
-      <a href="{$site_root}/ads/destroy?id={$ad->id}">Rimuovi</a>
+      <a data-confirm href="{$site_root}/ads/destroy?id={$ad->id}">Rimuovi</a>
       <a href="{$site_root}/ads/edit?id={$ad->id}">Modifica</a>
     {/foreach}
   </section>
